配置修改
