+++ source/src/main/java/io/mycat/mycat2/MycatCore.java
@@ -25,6 +25,8 @@ package io.mycat.mycat2;
 
 import java.io.IOException;
 
+import io.mycat.mycat2.beans.conf.ProxyConfig;
+import io.mycat.proxy.ConfigEnum;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -37,22 +39,15 @@ import io.mycat.util.YamlUtil;
  */
 public class MycatCore {
 	private static final Logger logger = LoggerFactory.getLogger(MycatCore.class);
-	public static final String MOCK_HOSTNAME = "host1";
-
-	public static final String MOCK_SCHEMA = "mysql";
 
 	public static void main(String[] args) throws IOException {
-		String mySeq = "1";
-		if (args.length > 0) {
-			mySeq = args[0];
-		}
-		String mycatConf = "mycat" + mySeq + ".yml";
-		logger.debug("load config for {}", mycatConf);
 		// mycat.conf的加载不需要在集群内
-		MycatConfig conf = YamlUtil.load(mycatConf, MycatConfig.class);
+		ProxyConfig proxy = YamlUtil.load(ConfigEnum.PROXY.getFileName(), ProxyConfig.class);
+		logger.debug("load config for {}", ConfigEnum.PROXY.getFileName());
+		MycatConfig conf = new MycatConfig(proxy);
 
 		ProxyRuntime runtime = ProxyRuntime.INSTANCE;
-		runtime.setProxyConfig(conf);
+		runtime.setConfig(conf);
 
 		int cpus = Runtime.getRuntime().availableProcessors();
 //	    int cpus = 1;
