--- source/src/main/java/io/mycat/mycat2/beans/DNBean.java
@@ -1,87 +1,0 @@
-/*
- * Copyright (c) 2016, OpenCloudDB/MyCAT and/or its affiliates. All rights reserved.
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
- *
- * This code is free software;Designed and Developed mainly by many Chinese
- * opensource volunteers. you can redistribute it and/or modify it under the
- * terms of the GNU General Public License version 2 only, as published by the
- * Free Software Foundation.
- *
- * This code is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
- * version 2 for more details (a copy is included in the LICENSE file that
- * accompanied this code).
- *
- * You should have received a copy of the GNU General Public License version
- * 2 along with this work; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- * Any questions about this component can be directed to it's project Web address
- * https://code.google.com/p/opencloudb/.
- *
- */
-package io.mycat.mycat2.beans;
-
-/**
- * refer a database on a mysql replica
- *
- * @author wuzhihui
- */
-
-public class DNBean {
-    private String database;
-    private String mysqlReplica;
-
-    public String getDatabase() {
-        return database;
-    }
-
-    public void setDatabase(String database) {
-        this.database = database;
-    }
-
-    public String getMysqlReplica() {
-        return mysqlReplica;
-    }
-
-    public void setMysqlReplica(String mysqlReplica) {
-        this.mysqlReplica = mysqlReplica;
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((database == null) ? 0 : database.hashCode());
-        result = prime * result + ((mysqlReplica == null) ? 0 : mysqlReplica.hashCode());
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        DNBean other = (DNBean) obj;
-        if (database == null) {
-            if (other.database != null)
-                return false;
-        } else if (!database.equals(other.database))
-            return false;
-        if (mysqlReplica == null) {
-            if (other.mysqlReplica != null)
-                return false;
-        } else if (!mysqlReplica.equals(other.mysqlReplica))
-            return false;
-        return true;
-    }
-
-    @Override
-    public String toString() {
-        return "DNBean [database=" + database + ", mysqlReplica=" + mysqlReplica + "]";
-    }
-}
