--- source/src/main/java/io/mycat/mycat2/beans/ShardingRuleBean.java
@@ -1,75 +1,0 @@
-/*
- * Copyright (c) 2016, OpenCloudDB/MyCAT and/or its affiliates. All rights reserved.
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
- *
- * This code is free software;Designed and Developed mainly by many Chinese
- * opensource volunteers. you can redistribute it and/or modify it under the
- * terms of the GNU General Public License version 2 only, as published by the
- * Free Software Foundation.
- *
- * This code is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
- * version 2 for more details (a copy is included in the LICENSE file that
- * accompanied this code).
- *
- * You should have received a copy of the GNU General Public License version
- * 2 along with this work; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- * Any questions about this component can be directed to it's project Web address
- * https://code.google.com/p/opencloudb/.
- *
- */
-package io.mycat.mycat2.beans;
-
-import java.util.Map;
-
-/**
- * sharding rule bean
- *
- * @author wuzhihui
- */
-public class ShardingRuleBean {
-    private String name;
-    private String algorithm;
-    private Map<String, String> params;
-
-    public ShardingRuleBean(String name, String algorithm, Map<String, String> params) {
-        super();
-        this.name = name;
-        this.algorithm = algorithm;
-        this.params = params;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getAlgorithm() {
-        return algorithm;
-    }
-
-    public void setAlgorithm(String algorithm) {
-        this.algorithm = algorithm;
-    }
-
-    public Map<String, String> getParams() {
-        return params;
-    }
-
-    public void setParams(Map<String, String> params) {
-        this.params = params;
-    }
-
-    @Override
-    public String toString() {
-        return "ShardingRuleBean [name=" + name + ", algorithm=" + algorithm + ", params=" + params + "]";
-    }
-
-
-}
