+++ source/src/main/java/io/mycat/mycat2/beans/conf/ClusterBean.java
@@ -1,9 +1,9 @@
-package io.mycat.mycat2.beans;
+package io.mycat.mycat2.beans.conf;
 
 /**
- * Desc: 对应mycat.yml文件中的cluster
+ * Desc: 集群配置类
  *
- * @date: 19/09/2017
+ * @date: 24/09/2017
  * @author: gaozhiwen
  */
 public class ClusterBean {
@@ -63,5 +63,11 @@ public class ClusterBean {
 
     public void setPrepareDelaySeconds(int prepareDelaySeconds) {
         this.prepareDelaySeconds = prepareDelaySeconds;
+    }
+
+    @Override
+    public String toString() {
+        return "ClusterBean{" + "enable=" + enable + ", ip='" + ip + '\'' + ", port=" + port + ", myNodeId='" + myNodeId + '\'' + ", allNodes='"
+                + allNodes + '\'' + ", prepareDelaySeconds=" + prepareDelaySeconds + '}';
     }
 }
