+++ source/src/main/java/io/mycat/mycat2/beans/conf/ProxyBean.java
@@ -1,9 +1,9 @@
-package io.mycat.mycat2.beans;
+package io.mycat.mycat2.beans.conf;
 
 /**
- * Desc: 对应mycat.yml文件中的proxy
+ * Desc: mycat代理配置类
  *
- * @date: 19/09/2017
+ * @date: 24/09/2017
  * @author: gaozhiwen
  */
 public class ProxyBean {
@@ -30,5 +30,10 @@ public class ProxyBean {
 
     public void setPort(int port) {
         this.port = port;
+    }
+
+    @Override
+    public String toString() {
+        return "ProxyBean{" + "ip='" + ip + '\'' + ", port=" + port + '}';
     }
 }
