+++ source/src/main/java/io/mycat/mycat2/beans/conf/ReplicaIndexConfig.java
@@ -1,4 +1,4 @@
-package io.mycat.mycat2.beans;
+package io.mycat.mycat2.beans.conf;
 
 import io.mycat.proxy.Configurable;
 
@@ -10,7 +10,7 @@ import java.util.Map;
  * @date: 10/09/2017
  * @author: gaozhiwen
  */
-public class ReplicaIndexBean implements Configurable {
+public class ReplicaIndexConfig implements Configurable {
     private Map<String, Integer> replicaIndexes;
 
     public Map<String, Integer> getReplicaIndexes() {
