+++ source/src/main/java/io/mycat/proxy/ConfigEnum.java
@@ -1,9 +1,6 @@
 package io.mycat.proxy;
 
-import io.mycat.mycat2.beans.ReplicaConfBean;
-import io.mycat.mycat2.beans.ReplicaIndexBean;
-import io.mycat.mycat2.beans.SchemaConfBean;
-import io.mycat.mycat2.beans.ShardingRuleBean;
+import io.mycat.mycat2.beans.conf.*;
 
 import java.util.stream.Stream;
 
@@ -14,10 +11,14 @@ import java.util.stream.Stream;
  * @author: gaozhiwen
  */
 public enum ConfigEnum {
-    DATASOURCE((byte) 1, "datasource.yml", ReplicaConfBean.class),
-    REPLICA_INDEX((byte) 2, "replica-index.yml", ReplicaIndexBean.class),
-    SCHEMA((byte) 3, "schema.yml", SchemaConfBean.class),
-    SHARDING_RULE((byte) 4, "sharding-rule.yml", ShardingRuleBean.class);
+    PROXY((byte) 1, "mycat.yml", ProxyConfig.class),
+    CLUSTER((byte) 2, "cluster.yml", ClusterConfig.class),
+    BALANCER((byte) 3, "balancer.yml", BalancerConfig.class),
+    HEARTBEAT((byte) 4, "heartbeat.yml", HeartbeatConfig.class),
+    DATASOURCE((byte) 5, "datasource.yml", DatasourceConfig.class),
+    REPLICA_INDEX((byte) 6, "replica-index.yml", ReplicaIndexConfig.class),
+    SCHEMA((byte) 7, "schema.yml", SchemaConfig.class),
+    SHARDING_RULE((byte) 8, "sharding-rule.yml", ShardingRuleConfig.class);
 
     private byte type;
     private String fileName;
@@ -45,6 +46,4 @@ public enum ConfigEnum {
         return Stream.of(ConfigEnum.values()).filter(configEnum -> configEnum.type == type)
                 .findFirst().orElse(null);
     }
-
-    public static final int INIT_VERSION = 1;
 }
