+++ source/src/main/java/io/mycat/proxy/ProxyRuntime.java
@@ -16,6 +16,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 import io.mycat.mycat2.ConfigLoader;
 import io.mycat.mycat2.beans.ReplicaIndexBean;
+import io.mycat.mycat2.beans.conf.ProxyConfig;
 import io.mycat.proxy.man.cmds.ConfigUpdatePacketCommand;
 import io.mycat.util.YamlUtil;
 import org.slf4j.Logger;
@@ -38,9 +39,9 @@ import io.mycat.proxy.man.MyCluster;
 import io.mycat.util.TimeUtil;
 
 public class ProxyRuntime {
-	
+	public static final ProxyRuntime INSTANCE = new ProxyRuntime();
 	private static final Logger logger = LoggerFactory.getLogger(ProxyRuntime.class);
-	
+
 	/*
 	 * 时间更新周期
 	 */
@@ -49,9 +50,8 @@ public class ProxyRuntime {
 	private static final String PROCESSOR_CHECK    = "PROCESSOR_CHECK";
 	private static final String REPLICA_ILDE_CHECK = "REPLICA_ILDE_CHECK";
 	private static final String REPLICA_HEARTBEAT  = "REPLICA_HEARTBEAT";
-	
-	private ProxyConfig proxyConfig;
-	public static final ProxyRuntime INSTANCE = new ProxyRuntime();
+
+	private MycatConfig config;
 	private AtomicInteger sessionId = new AtomicInteger(1);
 	private int nioReactorThreads = 2;
 	private boolean traceProtocol = false;
@@ -98,7 +98,7 @@ public class ProxyRuntime {
 	public void init() {
 		//心跳调度独立出来，避免被其他任务影响
 		heartbeatScheduler = Executors.newSingleThreadScheduledExecutor();
-		timerExecutor = ExecutorUtil.create("Timer", getProxyConfig().getHeartbeat().getTimerExecutor());
+		timerExecutor = ExecutorUtil.create("Timer", getConfig().getHeartbeat().getTimerExecutor());
 		businessExecutor = ExecutorUtil.create("BusinessExecutor",Runtime.getRuntime().availableProcessors());
 		listeningExecutorService = MoreExecutors.listeningDecorator(businessExecutor);
 	}
@@ -195,12 +195,12 @@ public class ProxyRuntime {
 		this.myCLuster = myCLuster;
 	}
 
-	public ProxyConfig getProxyConfig() {
-		return proxyConfig;
+	public MycatConfig getConfig() {
+		return config;
 	}
 
-	public void setProxyConfig(ProxyConfig proxyConfig) {
-		this.proxyConfig = proxyConfig;
+	public void setConfig(MycatConfig config) {
+		this.config = config;
 	}
 
 	public static ScheduledExecutorService getSchedulerservice() {
