+++ source/src/main/resources/cluster.yml
@@ -1,0 +1,7 @@
+cluster:
+  enable: false
+  ip: 0.0.0.0
+  port: 9066
+  myNodeId: leader-1
+  allNodes: leader-1:127.0.0.1:9066,leader-2:127.0.0.1:9067,leader-3:127.0.0.1:9068
+  prepareDelaySeconds: 30
