+++ source/src/main/resources/mycat.yml
@@ -1,0 +1,3 @@
+proxy:
+  ip: 0.0.0.0
+  port: 8066
