+++ source/src/main/resources/schema.yml
@@ -1,28 +1,28 @@
 schemas:
   - name: test
-    type: DB_IN_ONE_SERVER
+    schemaType: DB_IN_ONE_SERVER
     defaultDN:
       database: test
       mysqlReplica: test
 #  - name: test2
-#    type: DB_IN_MULTI_SERVER
+#    schemaType: DB_IN_MULTI_SERVER
 #    defaultDN:
 #      database: mysql
 #      mysqlReplica: mysql
 #    tables:
 #      - name: t1
-#        type: MASTER
+#        tableType: MASTER
 #        store: test1:test #db:rep
 #      - name: t2
-#        type: SLAVE
+#        tableType: SLAVE
 #        store: test1
 #  - name: test3
-#    type: SQL_PARSE_ROUTE
+#    schemaType: SQL_PARSE_ROUTE
 #    defaultDN:
 #      database: test
 #      mysqlReplica: test
 #    tables:
 #      - name: shard1
-#        type: SLAVE
+#        tableType: SLAVE
 #        shardingKey: id
 #        shardingRule: auto-sharding-1
